Added defer to JavaScript '<script>' tag
I added the 'defer' attribute to the '<script>' tag to ensure that the JavaScript file is loaded after HTML parsing is complete, but before the 'DOMContentLoaded' event occurs. By making this change, I was able to move the script loading to the '<head>' tag instead of keeping it at the end of the '<body>'. This modification provides a more intuitive structure for my code.